Additional split-screen fixes, enable video fixes and put in the input fixes for split-screen hack.
